requirements.txt
# Core ML and Vision Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.8.0
Pillow>=9.5.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Processing
pandas>=2.0.0
albumentations>=1.3.0
imageio>=2.31.0

# Web Interface
streamlit>=1.25.0
gradio>=3.40.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Jupyter and Visualization
jupyter>=1.0.0
jupyterlab>=4.0.0
plotly>=5.15.0
bokeh>=3.2.0

# Model Utilities
timm>=0.9.0
transformers>=4.30.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Optional GPU acceleration
# Uncomment if you have CUDA installed
# torch-audio>=2.0.0
# torchtext>=0.15.0